# docker/runpod-5090/openapi.yaml
openapi: 3.1.0
info:
  title: FLUX.1-dev Inference API
  description: |
    Image generation API powered by FLUX.1-dev running on RunPod with RTX 5090.
    Uses FP8 quantized flow transformer with bf16 text encoder and VAE.
    Supports LoRA loading/unloading for style and character adapters.
  version: 2.0.0

servers:
  - url: https://{pod_id}-8088.proxy.runpod.net
    description: RunPod proxy endpoint
    variables:
      pod_id:
        description: RunPod pod identifier
        default: REPLACE_ME

security:
  - bearerAuth: []

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer token set via AUTH_TOKEN environment variable on the pod

  schemas:
    GenerateRequest:
      type: object
      required:
        - prompt
      properties:
        prompt:
          type: string
          description: |
            Text description of the image to generate. Processed by T5 (up to 512 tokens) and CLIP (up to 77 tokens).
            Supports prompt weighting syntax:
            - `(word)` increases attention by 1.1x
            - `(word:1.5)` increases attention by 1.5x
            - `[word]` decreases attention by 1.1x
            - Nesting stacks: `((word))` = 1.21x
          example: "A (beautiful:1.3) sunset over a mountain landscape with ((dramatic lighting))"
        width:
          type: integer
          default: 720
          description: Image width in pixels. Keep total pixels at or below 1024x1024 equivalent to avoid OOM.
          example: 1024
        height:
          type: integer
          default: 1024
          description: Image height in pixels. Keep total pixels at or below 1024x1024 equivalent to avoid OOM.
          example: 1024
        num_steps:
          type: integer
          default: 24
          description: Number of diffusion steps. More steps = better quality but slower. Range 8-50 typical.
          example: 28
        guidance:
          type: number
          format: float
          default: 3.5
          description: Guidance scale. Higher values follow the prompt more closely. Range 1.0-10.0 typical.
          example: 3.5
        seed:
          type: integer
          minimum: 1
          maximum: 4294967295
          description: Random seed for reproducible generation. If omitted, a random seed is used.
          example: 42
        strength:
          type: number
          format: float
          default: 1.0
          minimum: 0.0
          maximum: 1.0
          description: Diffusion strength when using init_image. 1.0 = full generation, lower values preserve more of the init image.
        init_image:
          type: string
          description: Base64-encoded image or file path on the pod to use as the starting image for img2img generation.
        num_images:
          type: integer
          default: 1
          minimum: 1
          maximum: 4
          description: Number of images to generate in a single request.
          example: 1
        jpeg_quality:
          type: integer
          default: 99
          minimum: 1
          maximum: 100
          description: JPEG compression quality for output images.
          example: 99

    GenerateResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique request identifier.
        status:
          type: string
          enum: [success, error]
        input:
          type: object
          description: Echo of all input parameters used for generation (excluding init_image).
        output:
          type: array
          items:
            $ref: '#/components/schemas/GenerateOutput'
          description: List of generated images with download URLs.
        seed:
          type: integer
          description: The actual seed used for generation. Use this to reproduce the exact same image.
        metrics:
          type: object
          properties:
            generation_time_seconds:
              type: number
              format: float
              description: Total generation time in seconds.

    GenerateOutput:
      type: object
      properties:
        url:
          type: string
          description: Relative URL to download the generated image. Prepend the pod base URL to get the full path.
          example: "/images/a1b2c3d4-5678-90ab-cdef-1234567890ab_0.jpg"
        width:
          type: integer
          example: 1024
        height:
          type: integer
          example: 1024

    LoraRequest:
      type: object
      properties:
        path:
          type: string
          description: Absolute path to the LoRA safetensors file on the pod.
          example: "/workspace/loras/my-style-lora.safetensors"
        scale:
          type: number
          format: float
          default: 1.0
          description: Scaling factor for LoRA weights. 0.5-1.0 typical for style LoRAs.
          example: 0.7
        name:
          type: string
          description: Identifier for the LoRA. Used for unloading. If omitted, derived from filename.
          example: "mystyle"
        action:
          type: string
          enum:
            - load
            - unload
          default: load
          description: Whether to load or unload the LoRA.

    LoraResponse:
      type: object
      properties:
        status:
          type: string
          enum:
            - success
            - error
        message:
          type: string
          description: Error message if status is "error".

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Unauthorized"

paths:
  /generate:
    post:
      summary: Generate an image
      description: |
        Generates an image from a text prompt using FLUX.1-dev.

        **Default (format=json):** Returns JSON with generation parameters, seed, timing metrics,
        and temporary download URL(s) for the generated image(s). Images are available for 30 minutes.

        **Backward compat (format=stream):** Returns raw streaming JPEG bytes (legacy behavior).

        **Recommended resolutions:**
        - Square: 1024x1024
        - Landscape 16:9: 1280x720
        - Portrait 9:16: 720x1280
        - Portrait 3:4: 768x1024
        - Landscape 4:3: 1024x768
      operationId: generateImage
      security:
        - bearerAuth: []
      parameters:
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum: [json, stream]
            default: json
          description: Response format. "json" returns structured response with image URL. "stream" returns raw JPEG bytes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateRequest'
            examples:
              simple:
                summary: Basic generation
                value:
                  prompt: "A fluffy orange cat sitting on a windowsill at sunset"
                  width: 1024
                  height: 1024
                  num_steps: 28
              portrait:
                summary: Portrait with seed
                value:
                  prompt: "Professional headshot of a woman with curly red hair, soft studio lighting"
                  width: 720
                  height: 1280
                  num_steps: 28
                  seed: 12345
              batch:
                summary: Multiple images
                value:
                  prompt: "A cyberpunk city street at night, neon signs, rain"
                  width: 1024
                  height: 1024
                  num_steps: 28
                  num_images: 2
      responses:
        '200':
          description: Generation successful (format=json)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateResponse'
              example:
                id: "a1b2c3d4-5678-90ab-cdef-1234567890ab"
                status: "success"
                input:
                  prompt: "A fluffy orange cat sitting on a windowsill at sunset"
                  width: 1024
                  height: 1024
                  num_steps: 28
                  guidance: 3.5
                  seed: 42
                  strength: 1.0
                  num_images: 1
                  jpeg_quality: 99
                output:
                  - url: "/images/a1b2c3d4-5678-90ab-cdef-1234567890ab_0.jpg"
                    width: 1024
                    height: 1024
                seed: 42
                metrics:
                  generation_time_seconds: 4.2
            image/jpeg:
              schema:
                type: string
                format: binary
              description: Raw JPEG bytes (when format=stream)
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error (e.g. CUDA OOM)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /images/{image_id}:
    get:
      summary: Download a generated image
      description: |
        Serves a temporarily stored generated image. Images are automatically
        deleted after 30 minutes. The URL is returned in the POST /generate response.
      operationId: getImage
      security:
        - bearerAuth: []
      parameters:
        - name: image_id
          in: path
          required: true
          schema:
            type: string
          description: Image identifier (UUID_index.jpg format)
          example: "a1b2c3d4-5678-90ab-cdef-1234567890ab_0.jpg"
      responses:
        '200':
          description: Image file
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Image not found or expired
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Image not found or expired"

  /lora:
    post:
      summary: Load or unload a LoRA
      description: |
        Loads or unloads a LoRA checkpoint into the flow transformer.
        LoRAs are fused directly into model weights (no additional VRAM cost).
        The LoRA file must already exist on the pod filesystem.
      operationId: loraAction
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoraRequest'
            examples:
              load:
                summary: Load a style LoRA
                value:
                  path: "/workspace/loras/watercolor.safetensors"
                  scale: 0.7
                  name: "watercolor"
                  action: "load"
              unload:
                summary: Unload a LoRA
                value:
                  name: "watercolor"
                  action: "unload"
      responses:
        '200':
          description: LoRA action completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoraResponse'
        '400':
          description: Invalid action
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoraResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoraResponse'